backend/requirements.txt
APScheduler==3.10.4
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-filter==25.1
django-cors-headers==4.3.1
djangorestframework==3.16.0
djangorestframework-simplejwt>=5.2.2
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==25.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
